board_f: Make fdt_size an int
This shouldn't be above a few MB so use an int to save stack space on
64-bit machines.

+++ include/board_f.h
@@ -21,7 +21,7 @@ struct board_f {
 	/**
 	 * @fdt_size: space reserved for relocated device space
 	 */
-	unsigned long fdt_size;
+	unsigned int fdt_size;
 	/**
 	 * @new_bootstage: relocated boot stage information
 	 */
